requirements.txt
picamerax[array]
numpy
omegaconf
flask
Pillow